# internal/apidocs/openapi.v1.yaml
openapi: 3.0.3
info:
  title: audistro-catalog API
  version: 1.0.0
  description: HTTP API for artists, assets, payees, providers, playback, reports, and dev ingest helpers.
servers:
  - url: /
paths:
...
  /v1/admin/bootstrap/artist:
    post:
      summary: Bootstrap artist and payee mapping in dev
      operationId: bootstrapArtist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BootstrapArtistRequest'
      responses:
        '200':
          description: Artist and payee mapping created or reused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BootstrapArtistResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorResponse'
...
components:
...
  schemas:
    APIError:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
    APIErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/APIError'
...
    BootstrapPayeeRequest:
      type: object
      additionalProperties: false
      required:
        - fap_public_base_url
        - fap_payee_id
      properties:
        payee_id:
          type: string
        fap_public_base_url:
          type: string
          format: uri
        fap_payee_id:
          type: string
    BootstrapArtistRequest:
      type: object
      additionalProperties: false
      required:
        - handle
        - display_name
        - payee
      properties:
        artist_id:
          type: string
        handle:
          type: string
        display_name:
          type: string
        pubkey_hex:
          type: string
        payee:
          $ref: '#/components/schemas/BootstrapPayeeRequest'
    BootstrapArtistResponse:
      type: object
      additionalProperties: false
      required:
        - artist_id
        - payee_id
        - handle
        - fap_payee_id
      properties:
        artist_id:
          type: string
        payee_id:
          type: string
        handle:
          type: string
        fap_payee_id:
          type: string
...